cleanup of elaboration
