+++ manual/trees/0002.tree
@@ -72,8 +72,4 @@
 
 \transclude{000B}
 
-\subtree[000I]{
-  \title{Organization}
-
-  \p{}
-}
+\transclude{000I}
