+++ packages/coln-compiler/src/Coln/Core/Memoed.hs
@@ -94,7 +94,7 @@ instance (V.HasEvaluation c) => LevelOf (Ty c) where
 instance Readback (Memoed a b c) (a c) where
   readb _ m = m.stx
 
-mkGlobal :: Name -> V.Ty N -> El D -> GlobalEntry
-mkGlobal n ty x = do
+mkGlobal :: Name -> V.Ty N -> El D -> Mode -> GlobalEntry
+mkGlobal n ty x m = do
   let neu = V.reflect (V.GlobalVar n neu) V.Id ty (Just x.val)
-  GlobalEntry x.stx neu ty
+  GlobalEntry x.stx neu ty m
